Amalgamated test projects
