+++ BadMelon.API.sln
@@ -7,9 +7,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BadMelon.API", "src\API\Bad
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BadMelon.Data", "src\Data\BadMelon.Data.csproj", "{043B31FE-4979-474F-A732-216A3C86EA0B}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BadMelon.Tests.Data", "tests\Data\BadMelon.Tests.Data.csproj", "{1AA2E8ED-5D9C-4204-9157-EDA1D970BD9E}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BadMelon.Tests.API", "tests\API\BadMelon.Tests.API.csproj", "{BD36CE5A-F153-4F93-B04C-A89003EB5CA4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BadMelon.Tests", "tests\BadMelon.Tests.csproj", "{BD36CE5A-F153-4F93-B04C-A89003EB5CA4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,10 +23,6 @@ Global
 		{043B31FE-4979-474F-A732-216A3C86EA0B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{043B31FE-4979-474F-A732-216A3C86EA0B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{043B31FE-4979-474F-A732-216A3C86EA0B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1AA2E8ED-5D9C-4204-9157-EDA1D970BD9E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1AA2E8ED-5D9C-4204-9157-EDA1D970BD9E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1AA2E8ED-5D9C-4204-9157-EDA1D970BD9E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1AA2E8ED-5D9C-4204-9157-EDA1D970BD9E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BD36CE5A-F153-4F93-B04C-A89003EB5CA4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BD36CE5A-F153-4F93-B04C-A89003EB5CA4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BD36CE5A-F153-4F93-B04C-A89003EB5CA4}.Release|Any CPU.ActiveCfg = Release|Any CPU
