+++ src/Data/Services/IRecipeService.cs
@@ -14,8 +14,6 @@ namespace BadMelon.Data.Services
 
         Task<Recipe> AddIngredientToRecipe(Guid recipeId, Ingredient ingredient);
 
-        Task<Recipe> AddIngredientToRecipe(Recipe recipe, Ingredient ingredient);
-
         Task<Recipe> UpdateIngredientInRecipe(Guid recipeId, Ingredient ingredient);
 
         Task<Recipe> DeleteIngredientInRecipe(Guid recipeId, Ingredient ingredient);
