+++ tests/Data/BadIngredientTypesTestData.cs
@@ -2,7 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 
-namespace BadMelon.Tests.API.Data
+namespace BadMelon.Tests
 {
     internal class BadIngredientTypesTestData : IEnumerable<object[]>
     {
